requirements.txt
PyQt5>=5.15.0
PyOpenGL>=3.1.5
PyOpenGL-accelerate>=3.1.5
numpy>=1.21.0
